+++ demisto_sdk/__main__.py
@@ -653,7 +653,14 @@ def generate_test_playbook(**kwargs):
     help="The path to the zip file downloaded via the Marketplace contribution UI. "
     "This will format the contents of the zip file to a pack format ready for contribution "
     "in the content repository on your local machine. The command should be executed from the "
-    "content repository's base directory.")
+    "content repository's base directory. When this option is passed, the only other options "
+    "that are considered are \"name\" and \"description\" - all others are ignored.")
+@click.option(
+    '-d', '--description',
+    type=click.STRING,
+    default='',
+    help="The description to attach to the converted contribution pack. Used when the \"contribution\" "
+    "option is passed.")
 def init(**kwargs):
     initiator = Initiator(**kwargs)
     initiator.init()